+++ fernet_fields/fields.py
@@ -9,48 +9,46 @@ from . import hkdf
 
 
 __all__ = [
-    'EncryptedField',
-    'EncryptedTextField',
-    'EncryptedCharField',
-    'EncryptedEmailField',
-    'EncryptedIntegerField',
-    'EncryptedDateField',
-    'EncryptedDateTimeField',
+    "EncryptedField",
+    "EncryptedTextField",
+    "EncryptedCharField",
+    "EncryptedEmailField",
+    "EncryptedIntegerField",
+    "EncryptedDateField",
+    "EncryptedDateTimeField",
 ]
 
 
 class EncryptedField(models.Field):
     """A field that encrypts values using Fernet symmetric encryption."""
-    _internal_type = 'BinaryField'
+
+    _internal_type = "BinaryField"
 
     def __init__(self, *args, **kwargs):
-        if kwargs.get('primary_key'):
+        if kwargs.get("primary_key"):
             raise ImproperlyConfigured(
-                "%s does not support primary_key=True."
-                % self.__class__.__name__
+                "%s does not support primary_key=True." % self.__class__.__name__
             )
-        if kwargs.get('unique'):
+        if kwargs.get("unique"):
             raise ImproperlyConfigured(
-                "%s does not support unique=True."
-                % self.__class__.__name__
+                "%s does not support unique=True." % self.__class__.__name__
             )
-        if kwargs.get('db_index'):
+        if kwargs.get("db_index"):
             raise ImproperlyConfigured(
-                "%s does not support db_index=True."
-                % self.__class__.__name__
+                "%s does not support db_index=True." % self.__class__.__name__
             )
         super(EncryptedField, self).__init__(*args, **kwargs)
 
     @cached_property
     def keys(self):
-        keys = getattr(settings, 'FERNET_KEYS', None)
+        keys = getattr(settings, "FERNET_KEYS", None)
         if keys is None:
             keys = [settings.SECRET_KEY]
         return keys
 
     @cached_property
     def fernet_keys(self):
-        if getattr(settings, 'FERNET_USE_HKDF', True):
+        if getattr(settings, "FERNET_USE_HKDF", True):
             return [hkdf.derive_fernet_key(k) for k in self.keys]
         return self.keys
 
@@ -64,9 +62,7 @@ class EncryptedField(models.Field):
         return self._internal_type
 
     def get_db_prep_save(self, value, connection):
-        value = super(
-            EncryptedField, self
-        ).get_db_prep_save(value, connection)
+        value = super(EncryptedField, self).get_db_prep_save(value, connection)
         if value is not None:
             retval = self.fernet.encrypt(force_bytes(value))
             return connection.Database.Binary(retval)
@@ -81,28 +77,31 @@ class EncryptedField(models.Field):
         # Temporarily pretend to be whatever type of field we're masquerading
         # as, for purposes of constructing validators (needed for
         # IntegerField and subclasses).
-        self.__dict__['_internal_type'] = super(
+        self.__dict__["_internal_type"] = super(
             EncryptedField, self
         ).get_internal_type()
         try:
             return super(EncryptedField, self).validators
         finally:
-            del self.__dict__['_internal_type']
+            del self.__dict__["_internal_type"]
 
 
 def get_prep_lookup(self):
     """Raise errors for unsupported lookups"""
-    raise FieldError("{} '{}' does not support lookups".format(
-        self.lhs.field.__class__.__name__, self.lookup_name))
+    raise FieldError(
+        "{} '{}' does not support lookups".format(
+            self.lhs.field.__class__.__name__, self.lookup_name
+        )
+    )
 
 
 # Register all field lookups (except 'isnull') to our handler
 for name, lookup in models.Field.class_lookups.items():
     # Dynamically create classes that inherit from the right lookups
-    if name != 'isnull':
-        lookup_class = type('EncryptedField' + name, (lookup,), {
-            'get_prep_lookup': get_prep_lookup
-        })
+    if name != "isnull":
+        lookup_class = type(
+            "EncryptedField" + name, (lookup,), {"get_prep_lookup": get_prep_lookup}
+        )
         EncryptedField.register_lookup(lookup_class)