+++ runtests.py
@@ -4,7 +4,6 @@ import pytest
 import os
 import sys
 
-os.environ.setdefault(
-    'DJANGO_SETTINGS_MODULE', 'fernet_fields.test.settings.sqlite')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "fernet_fields.test.settings.sqlite")
 
 sys.exit(pytest.main())